Initial UI Customization
While this is still a work in progress, a bulk of the initial UI has
been added.

The pipelines have been modified so that gulp is ran on every change to
any css/js/html. This comes in handy when developing the front end
and only adds a second or two to the rebuild process.