(#11) Add Concept of Blog Series
The concept of a blog series (or code series) has been added. This
includes a new page that can be navigated to `series` which includes
a list of each series that can be clicked in to.

If a blog post is in a series, the other posts that are part of the
series will be navigatable from each blog post.