(#11) Shift layout files to add ExtraHeaders section
Some of the base layout files have been moved around and placed in a centralized `_LayoutBase` file. This was done mostly to enable the usage of ExtraHeaders (updating of meta data), on an individual page. This was not possible with the current setup.

This also reduces the amount of partials used, so it is easier to track down files when needing to edit.

ExtraHeaders section was utilized on the Series page to modify the Title and Description.