(#9) Optimize Skillz section for mobile
This makes the columns of the Skillz section on mobile larger, so that
the icons are not all squished and are readable. On desktop, everything
appears mostly how it was before.

In addition, support has been added to this section for dark mode so
that icons appear well on dark backgrounds as well as light.